+++ src/vector.cpp
@@ -15,7 +15,7 @@ void Vector::allocateData()
   _data = new REAL[size];
 }
 
-Vector::Vector(const size_t size_, const REAL val) : _data(nullptr), size(size_)
+Vector::Vector(const unsigned size_, const REAL val) : _data(nullptr), size(size_)
 {
   allocateData();
   fill(val);
@@ -45,17 +45,17 @@ REAL Vector::nrm2() const
   return sqrt((*this) * (*this));
 }
 
-size_t Vector::getSize() const
+unsigned Vector::getSize() const
 {
   return size;
 }
 
-REAL Vector::at(size_t i) const
+REAL Vector::at(unsigned i) const
 {
   return _data[i];
 }
 
-REAL& Vector::at(size_t i)
+REAL& Vector::at(unsigned i)
 {
   return _data[i];
 }
@@ -66,7 +66,7 @@ void Vector::print(const std::string& descr) const
   std::cout << descr << std::fixed << std::setprecision(1) << std::setw(7) 
 	    << std::left<< std::endl;;
 
-  for (size_t i = 0; i < size; i++)
+  for (unsigned i = 0; i < size; i++)
     {
       std::cout << "\t" << _data[i] << std::endl;
     }
@@ -76,7 +76,7 @@ bool operator==(const Vector & v1, const Vector & v2)
 {
   assert(v1.getSize() == v2.getSize());
 
-  for (size_t i = 0; i < v1.getSize(); i++)
+  for (unsigned i = 0; i < v1.getSize(); i++)
     {
       if (v1.at(i) != v2.at(i))
 	{
@@ -93,7 +93,7 @@ Vector operator*(REAL a, const Vector & v)
 
   Vector r{v.getSize()};
 
-  for (size_t i = 0; i < v.getSize(); i++)
+  for (unsigned i = 0; i < v.getSize(); i++)
     {
       r.at(i) = v.at(i) * a;
     }
@@ -108,7 +108,7 @@ REAL operator*(const Vector & v1, const Vector & v2)
 
   REAL sum = 0;
 
-  for (size_t i = 0; i < v1.getSize(); i++)
+  for (unsigned i = 0; i < v1.getSize(); i++)
     {
       sum += v1.at(i) * v2.at(i);
     }
@@ -122,7 +122,7 @@ Vector operator+(const Vector & v1, const Vector & v2)
 
   Vector r{v1.getSize()};
 
-  for (size_t i = 0; i < v1.getSize(); i++)
+  for (unsigned i = 0; i < v1.getSize(); i++)
     {
       r.at(i) = v1.at(i) + v2.at(i);
     }
